chore #131 : scrollToSection함수 공통 유틸 함수로 분리
