+++ src/common/components/layout/header-buttons/landing-middle-buttons.tsx
@@ -12,25 +12,6 @@ gsap.registerPlugin(ScrollToPlugin);
 function LandingMiddleButtons() {
   const theme = useTheme();
 
-  const scrollToSection = (id: string) => {
-    const element = document.getElementById(id);
-    const projectsContainer = document.getElementById('projects-container');
-
-    if (element) {
-      if (id === 'help-section') {
-        // 기존 scrollIntoView 메서드로는 중간 가로 스크롤 영역 때문에 제대로 동작하지 않음.
-        gsap.to(window, {
-          duration: 1,
-          scrollTo: {
-            y: element.offsetTop + (projectsContainer?.clientWidth || 0) * 4,
-          },
-        });
-      } else {
-        element.scrollIntoView({ behavior: 'smooth', block: 'start' });
-      }
-    }
-  };
-
   return (
     <div css={styles.container}>
       <Button
